Begin port of Define and new Workspace page
This is very large commit with scattered files since it became
unwieldy to manage all the changes over an extended time. The
compiled and optimized files will be removed in a later release
to clean up the source tree. Tagged releases will contain only
the built and optimized files.

Many of the JavaScript files need to be cleaned up or removed
since they are no longer used.