Explicitly reset collections on session open
This is necessary due to a change in Backbone 1.0.0+ where collections
will be updated in-place by default (and thus not reset). This ensures
the collection is reset and the 'reset' event fires.

This fixes the issue that caused the collection never to resolve
itself because the 'reset' event did not fire.