Refactor where UI context-based actions are available
For field-level contexts, the UI components have been moved from
FieldControl to FieldForm. This improves FieldControl by focusing
the view on operating on the condition operators and values.

Actions have been added to ConcepForm which operates at the
context-level.

The FieldForm and ConcepForm implements should remain consistent
(and will most likely be broken out into a base class) to enable
interchangeable use of the two. The concept-level context _may_
contain more than one field, but the ways it add/updates/removes
the condition is transparent.