Clean up and simplify logic for context nodes
The `create` option on fetch now must be a node type of either
'branch', 'condition' or 'composite' to choose the correct constructor.

Fix missing call to root.save() prior to saving the ContextModel itself.
This prevented the publicAttributes object from every getting updated
before being sent to the server.

Add CONTEXT_SAVE channel to transparently request the session be saved.