Pass view to header cells and sort on header click.
This change enables sorting for columns in the result view. Clicking on
the column header will cause the following sort changes depending on the
current sort status of the table:
    - If the table is already sorted ascending by the clicked column,
      then the table will be sorted descending by the clicked column.
    - If the table is already sorted desceding by the clicked column,
      then the table will be unsorted.
    - If the table is unsorted or sorted by another column, then the
      table will be sorted ascending by the clicked column.
Each time a column is clicked, the view is updated and saved which, in
turn, causes the table itself to redraw in the correct sort order.