Update FieldForm to render arbitrary number of controls
This now uses a CollectionView to render a series of controls given
based on the FieldControlOptions model. FieldForm now has the
addControl method which takes a Control view class and options which
will auto-render the control.

A side effect to this is that the designated chart area has been
removed in the field form template since it renders with the
.controls-region.

Code related to the 'managed' option has been removed. This existed
only to be symmetrical to the ConceptForm, but served no utility in
practice so it has been removed to clarity.