Use checkboxes for export types to allow multiple exports.
The downloads are each triggered by setting window.location. This is not
an ideal way of doing this but it is less difficult(horrendous?) than
dynamically inserting then removing an iframe on the page for each of
the export type downloads. For now, setting window.location and an
over-estimated artificial delay will do.