Add export capability in both monitored and unmonitored mode.
The export mode used depends on the version returned from the export
endpoint on the server. If the version is new enough then we will
monitor downloads(check for a cookie) and be able to tell right away if
the download complete normally or in an error state. If there is not
support on the server for notifications in the response(notification =
cookie) then we call the downloads in a more serial manner and rely on a
simple timeout before making a one off judgement on the status of the
download. While the unmonitored approach certainly has some drawbacks,
it is the only way to make this backwards compatible with older versions
of Serrano that don't set a cookie in the response. The serial download
code is only used as a last resort when the export endpoint doesn't
support notifications on export completion.