Refactor and simplification of context nodes
There is a known bug when viewing non-applied filters and them appearing
in the context panel. The behavior can be produced in two ways: 1. when
no filters are applied and the first time a concept is viewed, the help
text in the context panel disappears; 2. when a filter is applied,
removed, then changed (still not yet applied), the old filter will
appear in the panel.

This is due to the underlying collection of nodes being 'watched' which
does not take in account whether the filter is applied.