Fix case where clicking non-routeable link broke back button
For whatever reason, using Backbone.history.navigate to check if a route exists then calling preventDefault() was preventing a valid entry from being written to the browser's history. Using a separate method for checking for route existence then following the preventDefault and navigate steps resolves the problem. I have tested and this is no longer an issue in FF, IE, Chrome, or Safari.