Ports Makefile functionality to Grunt.
Previously the combination of js files and the symlinking of templates to allow for processing by requirejs was handled by a task in the Makefile. This was Cilantro's last remaining dependency on the Makefile to compile for distribution. This commit removes the dependency on the Makefile by porting the "combine" directive into "copy" and "symlink". Important to note is that this commit does not handle the existence of an extension directory.

This commit also adds a new Grunt task: buildDist which should compile coffeescript, sass, combine js, symlink templates, and build min and src distributions through requirejs into the dist folder.