Implementation of ContextManager
The manager maintains two tree structures containing context nodes.
The `working` tree contains local state and is never overridden by a
sync with a server (since a change could be made prior to the response
returning). The `upstream` tree contains the state last synced with the
server. It contains the last known stable (and applied) state and is
read-only.

The manager exposes methods for defining nodes and facilities for
comparing the working and upstream trees. The nodes themselves provide
methods for checking and manipulating their state. A node is considered
to be _managed_ if a manager has been defined locally for that node.

Nodes now provides mechanisms for declaring their state including:

- isNew
- isValid
- isDirty
- isRemoved
- isEnabled

As well as corresponding events:

- enable
- remove
- change

Branch nodes have been modified to be aware of their child nodes if the
`deep` option has been passed and is true.

Fix #183, fix #184