Simplify version comparison and add c.isSupported()
c.isSupported([minVersion]) returns a boolean as to whether the current
session's Serrano version is within range as defined by Cilantro's
minSerranoVersion and (now added) maxSerranoVersion. An alternate
minVersion and maxVersion can be passed to isSupported() to denote
Cilantro has support for a Serrano feature introduced at a particular
version and will expose it if the active Serrano version being used is
equal or exceeds the specified version.

See the cilantro/utils/version module for details