Give preference to show/hide filter button clicks
This commit respects the button click over the scroll action. That is,
if the user chooses to hide the filters by clicking the button then the
scroll action has no effect on the filter panel and only
expands/collapses the toolbar. If the user has take no action with the
button then the filter panel will show/hide on scroll as in the previous
commit.