Remove references of third-party libraries on cilantro object
This is to improve clarity of the code as well as make it possible
to load the individual third-party modules for use in the project or
override the core libraries if existing in a project.

The dropped references include:

- `c.$`
- `c._`
- `c.Backbone`
- `c.Marionette`
- `c.Highcharts`

These dependencies must now be included explicitly in the dependencies
array for the module.

In addition, the subscribe/publish hooks for Marionette.ItemView have
been removed. There were only two areas this was being used and they
were simply replaced with `c.mediator`.