Add filter button icon and hide toolbar button text on small displays
Bootstrap considers 1200px or more to be a large display. It just so
happens that our toolbar's layout gets out of whack right around that
dimension so the new CSS in the media selector deems a large display to
be 1200px and hides the button text in the navbar buttons when the
screen width is smaller than that.