Add strict option for BranchNodeMode.toJSON()
If strict is true, non-valid children will be excluded from the output.
If no children result, undefined is return. One additional change includes
not returning children that have falsy attributes, e.g. null/undefined.
This is used in combination with the strict option which prevents nested
empty branches from being returned. This is currently used when a working
node is applied upstream.