Propagate clear and change events to and from BranchNodeModel.children
The 'clear' event is triggered top-down which has the behavior of clearing
the contents of all nodes. This only _actually_ clears the upstream nodes,
but the working nodes still must receive the event to behave consistently.

Changes made to the children collection do not automatically update the
branch node's internal children attribute. The 'change' event is now
bound on children to silently update the internal attributes to keep them
in sync. This is necessary for accurate comparison during branch change
events.