Hide results table when there are no results
A while ago, there was a problem where no results left a loading
indicator spinning away in the results table. There was an attempt to
fix this(removed in this commit) but that fix suffered from leaving the
'no data' message showing when paginating. Rather than continue trying
to jam a message into the table, this commit simply hides the whole
table when there are no results. The user can clearly see that there are
0 results(shown in paginator and context panel) so there is no need to
add a third indicator in the table and show table headers without any
actual data.