Handle case when resize occurs when context panel is not visible
When the browser was resized while on the query page, navigating to
the results view resulted in the context panel offset from the right.
This grossly fixes the issue by binding to the 'router:load' event to
'show' the context panel which handles adjusting the positioning.

This is a temporary fix until the context panel is refactored in #258