Increase the timeouts for monitored and unmonitored exports
For unmonitored exports, all we do is see if an error occurs within the timeout window, if not, we wait until the end of the window and then inform(lie to)? the user that the export succeeded. In modern versions of Serrano that support monitored timeouts, if the request times out then we show a new timeout status to the user. NOTE: This does not abort the request because the damage will already be done on the server(as far as the database is concerned).