Remove global router in favor of session-based routers
This introduces a new module that acts as the 'default' module for
bootstrapping Cilantro with the default workflows. The `url` and
`credentials` in the configuration object will be used by this module
if included. This module also acts as a template for other main modules
to emulate after.

`c.session` and `c.router` will continue to reference the _current_ of
each when the session is successfully opened.