Make use ContextNodeModel.isValid for asserting validity
This previously required passing the `strict` option to only include
valid nodes in the output. However having a non-strict behavior has not
proven to be useful so it has been removed.

It was originally intended to allow serialization of the working nodes
to include non-valid nodes, but this is not being used anywhere in the
codebase.