Simplify names and fix logic for various cases
The `from-` and `to-` prefixes have been renamed to `upper-` and `lower-`
to be consistent with methods names.

setValue must always reset the values prior to setting new ones in
case an array-based value goes to a non-array base value (the other
value would not be cleared).

The local reference `@operator` has been removed since the order isn't
technically guarantee for the operator being set prior to the value.