Fix logic and simplify cilantro.isSupported
Previously the version passed set bounded the minimum version supported
which was wrong. This function has been changed to require a version to
be passed. The max version defaults to Cilantro's unless the API's version
is less than what Cilantro supports. For example, the following:

    cilantro.isSupported('2.1.2')

returns true if the current endpoint's version is at least 2.1.2. The
minSerranoVersion and maxSerranoVersion variables are simply used to
notify when Cilantro does not support (less than minimum) or has not been
tested (greater than maximum) the current version of Serrano.