Add editable flag to determine whether user can edit query
This flag is explicitly set to false for public queries as we don't want people to be able to edit them. Previously, I used the publicity of the query to determine whether the query should be editable but this had the side effect of making user-owned public queries uneditable now that Serrano is returning the publicity for owned and shared queries.