Handle case where query collections sync prior to workspace view init
Prior to this, if there were no queries in a collection and the user did not navigate to workspace/ as their first view then the loading indicators would never disappear because the empty message was only shown in response to the collection sync event which would occur prior to the workspace view being initialized. This is explained in the added code.