Update public queries on sync AND destroy events on query collection
This fixes the problem where public queries could be deleted but the list of public queries was not updated until the page was reloaded. Due to recent changes in the way the delete is handled, we need to subscribe to the destroy event in addition to the sync event to account for deletion of models in the query collection.