Remove '...' from show/hide filters button
This ends up saving about 40px of horizontal space in the toolbar which
should allow more room for the text in the count label on the left.
Doesn't seem like much but it is pretty useful in projects like varify
which might have long sample names after the record count. Personally, I
didn't care for the ellipsises either even though I probably added them
in the first place.