Promote Cilantro controls API to public namespace, `c.controls`
This fixes a bug that incorrectly mixed in the controls set/get
methods into the `ui` namespace. This also provides support for
loading modules remotely like the templates API.

The behavior described in #419 is not desirable since the controls
configuration options must be available when Cilantro loads as supposed
to when the session starts.

Close #419