Add option to toggle condensing multi-field concepts
The previous behavior was to render the non-initial field in a more
condensed representation. This tends to denote the non-initial fields are
less important the first one, which should not be the case. The fix
introduces a `ConceptForm` level option `condense` that when set to true
will use the legacy behavior of condensing the non-initial fields. By
default this option is false.

To use the legacy behavior, simply set the `concepts.defaults.form.condense`
to true to apply to all concepts.

Fix #360