Add config option for using fixed view on preview endpoint.
The view can be specified by setting the preview.view config option. An
example is shown below:

    preview:
        view:
            columns:[2]

In the above example, all calls to the preview endpoint will use a view
composed of only the concept with ID 2. This is an example taken from
varify to use only the sample column in calls to the preview endpoint.

Signed-off-by: Don Naegely <naegelyd@gmail.com>