Change configuration option for better namespacing
Session-based data are namespaced under `session.NS.data` where `NS` is either
`defaults` or a named session. Named session are not yet supported, but this
ensure forward compatibility of the options. Each key in `data` corresponds to
a resource available for that session, such as `preview`. The value corresponds
to the custom request options or data used when requesting the resource.

Signed-off-by: Byron Ruth <b@devel.io>