Implement new pagination modules for models and views
The PaginatedCollection model internalizes the pages and collects and
organizes the items as pages are fetched. From the outset, this makes
the collection consumable like any other collection.

The PaginatedItems view is the compliment view to the PaginatedCollection
and supports "infinite scrolling". This makes the collection behave as if
the items were all pre-fetched.

The PaginationSearch is simply a convenience view to handle toggle between
the initial URL and one with a query parameter.

This will eventually phase out the existing pagination facilities in
cilantro/models/paginator and cilantro/ui/paginator.

Signed-off-by: Byron Ruth <b@devel.io>