Refactor of workspace layouts and side panels
The context and concept panels are now defined independent of any
workflow. They are namespaced under `cilantro.panels` once the
session starts. Fix #258

To accommodate fixed position panels, the HTML of the workflows have
been simplified and removed of the *-fluid elements. Panels are by
default fixed in width and adjacent elements are positioned
appropriately.

The refactor removed the need for the jquery-stacked so it has been
removed.

By design, this always fixes an overflow issue with the new "jump to"
links. Fix #470.

Signed-off-by: Byron Ruth <b@devel.io>