Replace concept-level form controls with field controls
Each field now has it's own set of apply/update button to remove the
ambiguity when applying a filter. This relies on the new context filters
API and uses the filter node directly. The context is still available
for performing checks between the applied and "working" filter states.

Control validation has been integrated to make validation and diffing
the states easier to use.

Fix #498, #448

Signed-off-by: Byron Ruth <b@devel.io>