Implement new context and filter models
The `Filter` model represents a single context node that is arbitrarily
complex. It is bound to a concept, field, or concept-field. Filters are
managed by the `Filters` collection which is a flat container. It
handles serializing the context nodes into a representation supported by
the server.

Includes update to Backbone 1.1.2

The context node API was a blocker for upgrading due to the changed
behavior of looking up models by id in collections.

One noteable change is that URLs are no longer bound to to model or
collections during initialization.

This version also has native AMD support.

Fix #318, #328, #467

Signed-off-by: Byron Ruth <b@devel.io>