Force applied filters to be enabled
This prevents the case that an existing filter that is disabled gets
updated, but has no upstream effect since it's disabled. In addition,
the  attribute is stripped to prevent using and old attribute.

As noted in the comment, a better method for cleaning the attributes
prior to being applied should be implemented.

Signed-off-by: Byron Ruth <b@devel.io>