Refactor support for required filters
The implementation has changed to be more generic. The new configuration
option `filters.required` is an array of lookups, that is, objects
containing keys that can be used to match a filter. This is usually a
concept id, field id or both. The option `query.concepts.required` is
deprecated, but will remain until the 2.4 release.

The 'context:required' event has been deprecated in favor of using just
the 'context:invalid' event since it is more general and contains the same
information.

Signed-off-by: Byron Ruth <b@devel.io>