Simplify search input view and fix various issues
The bound event is now 'input' which is supported by all major browsers
including IE 9+. The 'input' event does not fire for non-printing
characters and only when the text value changes.

Extracting the `delay` option has been fixed to account for a `options`
function.

The bound `renderInputValue` handler exits if the view triggered the
'search' event on itself.

Signed-off-by: Byron Ruth <b@devel.io>