Implement control for filtering on free-text data
The interface is a single text input that parses the keywords and
constructs a set filters representing the search string. Although a
server-side alternative for parsing and filtering on the search string
would be more optimal, this approach was taken to prevent requiring the
server-side support.

There is the foundation for supporting a "preview list" of values that
match, however the Serrano API support is lacking in the current version.
Therefore the preview is disabled by default and requires at the least
the next version 2.3.5 to display the preview list.

Signed-off-by: Byron Ruth <b@devel.io>