Prevent exclusion from persisting in infograph
This fixes the bug where, upon applying exclusion by clicking the
exclude checkbox, the excluded style would remain on all selected bars
and the exclusion operator(-in) would be used even when the checkbox was
then unchecked. This was due to the events being triggered on the
collection when all the models were being updated. See notes in code for
more details on current approach.

Signed-off-by: Don Naegely <naegelyd@gmail.com>