Fix issues with range control being set while changing
This was caused because range overrides the set() method of the base control class. In the base class, the control checks the _changing flags and aborts the set if _changing is true. The range control was not doing that before this change so it was causing all sorts of odd bugs especially when modifying the operator via the dropdown box.

Signed-off-by: Don Naegely <naegelyd@gmail.com>