Fix infograph bar display to use label
The bars are rendered using the field's distribution resource data. This
contains the count and array of values that were grouped for the
distribution (by default 1). The labels are not present in this
response, so this "fix" will fetch the values on the field to create a
mapping from value to label before rendering the the bars.

The values requests uses a applies a zero limit parameter under the
assumption that since the field is being rendered as an Infograph control,
there are a small number of values.

Signed-off-by: Byron Ruth <b@devel.io>