Fix case where navigating first to results page broke save columns button
This was an issue of where the event handler was initialized in the
workspace workflow. Formerly, the event handler was initialized in the
initialize method but the ui elements aren't initialized yet at that
point so we were trying to call 'show()' on a string(because ui elements
are string still during initialize). By setting the event handler in
onRender we can be sure that the ui elements have been initialized and
the event handler will be valid.

Signed-off-by: Don Naegely <naegelyd@gmail.com>