Update selected columns when view is synced
This fixes the case where a saved query is opened which has different columns than the current view. Then, the user opens the Change Columns modal and the columns there are different than those in the results table. Clicking Cancel then opening the dialog again fixed this. Now, the dialog will detect that sort of change and respond accordingly.

Signed-off-by: Don Naegely <naegelyd@gmail.com>