Add silent option to value set calls to avoid revalidation per model
Without this change, the validate method of the search control is
triggered for every single model because it listens to the events on the
collection. We now set everything silently and then, once the ajax call
is complete, trigger a change event that the search control will "hear"
and cause itself to revalidate.

Signed-off-by: Don Naegely <naegelyd@gmail.com>