Silence set in search control setValue, trigger manual reset
This fixes that case where setting the value was extremely slow for large lists of values. What was happening before was each value in the list was triggering an event and ui.values.reloadText() was being called(among other event handlers) due to all the events being generated per model during the set call. Now, the set call is silenced and then we manually trigger a reset.

Signed-off-by: Don Naegely <naegelyd@gmail.com>