Update session close to only reset collections
This is important for the stats element on the session data. The stats
element is a Backbone Model not a Backbone Collection so it does not
have a `reset()` method. The `close()` method for the session now checks
for the precense of the reset method on the data items before calling reset.
This is noted in the code as well with a comment.

Signed-off-by: Don Naegely <naegelyd@gmail.com>