Fix case where stats counts loader was not being hidden
This was occurrign when navigating to the Workspace page via a navbar
link from the Query or Results page. This was happening because the code
that hid the loader used the 'ui' element which isn't available until
the Workspace is rendered so it won't work when the collection resets
and the Workspace page has not been shown yet. This fixes that case
without having to use raw jquery selectors to lookup the UI elements.

Signed-off-by: Don Naegely <naegelyd@gmail.com>