Add support for canceling on the results query and exports
When the loading overlap is shown, a button is now available to cancel
the query. This aborts the request and sends a DELETE request to the
endpoint to notify the server to stop processing of the results. Upon
canceling the query, a message is displayed with suggestions of how to
improve query if was taking too (and likely very complex).

Signed-off-by: Byron Ruth <b@devel.io>