Set collections and UI elements based on async setting
This PR adds a setting to Cilantro that will choose whether synchronous or asycnhronous endpoints are used. The setting also controls which type of collections and UI elements are used as there can be big differences in behavior between the synchronous and asynchronous versions. The async collections and UI elements are stubbed out here for completeness and will be implemented in upcomin work.

Signed-off-by: Don Naegely <naegelyd@gmail.com>