Implement dialog for scripted API call
The motivation for this change is to provide a starting pointing for
writing scripts to export results based on query developed in the
interface.

The change introduces a new dialog containing pre-fabricated API
calls using cURL and the Python requests library for JSON or CSV
formats. The dialog can be opened using a new button in the results
toolbar.

Fix #777

Signed-off-by: Byron Ruth <b@devel.io>